WIP Match with API

+++ Flextime.Sync/DataContracts/SummaryDataContract.cs
@@ -1,5 +1,5 @@
 namespace Inhill.Flextime.Sync.DataContracts;
 
-public record DayDataContract(DateOnly Date, uint Hash);
+public record DayDataContract(DateOnly Date, long Hash);
 
 public record SummaryDataContract(DayDataContract[] Items);
